refactor, delete settings_preview
