+++ pages/api/supabase/settings/update.tsx
@@ -7,14 +7,19 @@ export default async function handler(
 ) {
     try {
         const payload = req.body;
-        const { data, error } = await supabase
-            .from(process.env.SETTINGS_DB!)
-            .insert({
-                pictureUrl: payload.pictureUrl,
-                uid: payload.uid,
-            });
 
-        if (error) res.status(500).json(error);
+        const { error: deleteError } = await supabase
+            .from(process.env.SETTINGS_DB!)
+            .delete()
+            .neq('id', 0);
+
+        if (deleteError) res.status(500).json(deleteError);
+
+        const { data, error: insertError } = await supabase
+            .from(process.env.SETTINGS_DB!)
+            .insert(payload);
+
+        if (insertError) res.status(500).json(insertError);
 
         res.status(200).json(data);
     } catch (error) {
