--- src/components/ui/admin/settings/listSlides/types/index.ts
@@ -1,8 +1,0 @@
-export type Slide = {
-    id: number;
-    img?: string;
-    fileDownload?: string;
-    title?: string;
-    desc?: string;
-    active: boolean;
-};
