+++ src/components/widgets/admin/Settings.tsx
@@ -1,32 +1,29 @@
-import { Card, Flex, Spin } from 'antd';
+import { Card, Flex, Typography } from 'antd';
 import { Images } from '~types/images';
 import { ListSlides, PicturesUpload } from '~ui/admin/settings';
+import { Slide } from '~types/settings';
 
 export const Settings = ({
     images,
     slidesData,
-    isLoadingImages,
 }: {
     images?: Images;
-    slidesData: any;
-    isLoadingImages: boolean;
+    slidesData: Slide[];
 }) => {
     return (
         <>
             <Flex vertical gap={20} align="center">
                 <Card
-                    title="Images storage"
+                    title={
+                        <Typography.Title level={2}>
+                            Images storage
+                        </Typography.Title>
+                    }
                     style={{ maxWidth: '1365px', width: '100%' }}
                 >
-                    {isLoadingImages ? (
-                        <Flex align="center" justify="center">
-                            <Spin size="large" />
-                        </Flex>
-                    ) : (
-                        <Flex vertical gap={10}>
-                            <PicturesUpload images={images} />
-                        </Flex>
-                    )}
+                    <Flex vertical gap={10}>
+                        <PicturesUpload images={images} />
+                    </Flex>
                 </Card>
             </Flex>
             <ListSlides slidesData={slidesData} />
