+++ src/hooks/settings/index.ts
@@ -1,2 +1,3 @@
 export * from './useGetSettings';
 export * from './useUpdateSettings';
+export * from './useGetSettingsBase64';
