+++ src/hooks/settings/useGetSettingsBase64.tsx
@@ -2,12 +2,12 @@ import { useQuery } from '@tanstack/react-query';
 import api from '~lib/api.helper';
 import { Settings } from '~types/settings';
 
-export const useGetSettings = () => {
+export const useGetSettingsBase64 = () => {
     return useQuery({
-        queryKey: ['fetchSettings'],
+        queryKey: ['fetchSettingsBase64'],
         queryFn: async () => {
             const { data } = await api.get<Settings[]>(
-                'api/supabase/settings/get',
+                'api/supabase/settings/getWithBase64',
             );
 
             return data.map((item) => item.settings);
