--- src/hooks/settings_preview/index.ts
@@ -1,3 +1,0 @@
-export * from './useGetSettingsPreview';
-export * from './useUpdateSettingsPreview';
-export * from './useGetSettingsPreviewBase64';
