+++ src/types/settings.ts
@@ -1,5 +1,21 @@
 export type Settings = {
-    pictureUrl: string;
-    createdAt: Intl.DateTimeFormat;
-    uid: string;
+    settings: {
+        id: number;
+        img: string;
+        title: string;
+        desc: string;
+        active: boolean;
+        base64?: string;
+    };
+    created_at: Intl.DateTimeFormat;
+    id: number;
+};
+
+export type Slide = {
+    id: number;
+    img?: string;
+    fileDownload?: string;
+    title?: string;
+    desc?: string;
+    active: boolean;
 };
