--- src/types/settingsPreview.ts
@@ -1,12 +1,0 @@
-export type SettingsPreview = {
-    settings: {
-        id: number;
-        img: string;
-        title: string;
-        desc: string;
-        active: boolean;
-        base64?: string;
-    };
-    created_at: Intl.DateTimeFormat;
-    id: number;
-};
